requirements.txt
# FastAPI framework and ASGI server
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Database and ORM
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Authentication and password management
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic[email]==2.10.3
python-multipart==0.0.18
python-dotenv==1.0.1
python-decouple==3.8
unidecode==1.3.8
greenlet==3.1.1
bleach==6.3.0
redis==7.1.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==7.0.0
httpx==0.28.1
aiosqlite==0.20.0

# Code Quality
ruff==0.14.10
mypy==1.19.1
pre-commit==4.5.1

# Optional (if used in your project)
email-validator==2.2.0
apscheduler==3.10.4
itsdangerous==2.2.0
slowapi==0.1.9

# Template engine
jinja2==3.1.4
